qr_generator/requirements.txt
qrcode==7.4.2
Pillow==10.3.0
supabase==2.4.3
python-dotenv==1.0.1
streamlit==1.32.2
